# specs/001-mvp-closed-loop/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PPTagent MVP API
  version: 0.1.0-mvp
  description: |
    REST API for PPTagent MVP 业务闭环（生成 / 知识库 / Agent 进化）。
    与 WebSocket 事件契约（events.yaml）配合使用，错误码遵循 RFC 7807。
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev
  - url: https://api.pptagent.example.com/api/v1
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
paths:
...
  /generations:
    post:
      summary: 一句话生成 PPT
      operationId: createGeneration
      tags: [Generations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequest'
      responses:
        '202':
          description: 已入队（含 queue_position）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationTask'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /generations/{task_id}/trace:
    parameters:
      - in: path
        name: task_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 获取生成轨迹（FR-015）
      operationId: getTrace
      tags: [Generations]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TraceStage' }
        '404': { $ref: '#/components/responses/NotFound' }
  /generations/{task_id}/stages/{stage_name}/redo:
    parameters:
      - in: path
        name: task_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: stage_name
        required: true
        schema: { type: string, enum: [outline, points, svg, pptx] }
    post:
      summary: 重做指定阶段（FR-016）
      operationId: redoStage
      tags: [Generations]
      responses:
        '202': { description: 已入队 }
        '409': { $ref: '#/components/responses/Conflict' }

  # ─── Samples ───
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 形如 `Bearer <key>`（最小权限 IV + 速率限制）
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: 未认证
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 状态冲突（如取消已完成任务）
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: 并发超限 / 配额耗尽
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    CreateGenerationRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 1000
          description: 一句话需求（FR-001）
        referenced_sample_ids:
          type: array
          items: { type: string, format: uuid }
          description: 可选，指定本次任务使用的样本子集
    GenerationTask:
      type: object
      properties:
        id: { type: string, format: uuid }
        prompt: { type: string }
        status:
          type: string
          enum: [queued, running, success, failed, cancelled, archived]
        current_stage:
          type: string
          enum: [outline, points, svg, pptx, null]
        queue_position:
          type: integer
          nullable: true
          description: FR-029 排队位置
        result_pptx_url:
          type: string
          format: uri
          nullable: true
        style_fit_score:
          $ref: '#/components/schemas/StyleFitScore'
        token_consumed: { type: integer }
        estimated_tokens: { type: integer, nullable: true }
        estimated_seconds: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, nullable: true }
    TraceStage:
      type: object
      properties:
        id: { type: string, format: uuid }
        stage_name:
          type: string
          enum: [outline, points, svg, pptx]
        stage_order: { type: integer }
        input_summary: { type: string }
        output_summary: { type: string }
        referenced_sample_ids:
          type: array
          items: { type: string, format: uuid }
        duration_ms: { type: integer }
        status:
          type: string
          enum: [pending, running, success, failed]
        error_message: { type: string, nullable: true }
        redo_count: { type: integer }
...
    StyleFitScore:
      type: object
      properties:
        layout: { type: number, minimum: 0, maximum: 1 }
        palette: { type: number, minimum: 0, maximum: 1 }
        font: { type: number, minimum: 0, maximum: 1 }
        overall_hit: { type: boolean, description: FR-028 三层均 ≥ 80% 时为 true }
...
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: PPTagent 自定义错误码，详见 error-codes.yaml }
        request_id: { type: string, format: uuid, description: 关联 X-Request-Id }
        trace_id: { type: string, description: OTel trace ID }
        errors:
          type: array
          description: 字段级错误（422 验证失败时填充）
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }

    # ─── Pagination ───
...